+++ packages/kit/src/utils/routing.spec.js
@@ -1,6 +1,6 @@
 import { test } from 'uvu';
 import * as assert from 'uvu/assert';
-import { parse_route_id } from './routing.js';
+import { exec, parse_route_id } from './routing.js';
 
 const tests = {
 	'': {
@@ -25,6 +25,21 @@ const tests = {
 	},
 	'blog/[slug].json': {
 		pattern: /^\/blog\/([^/]+?)\.json$/,
+		names: ['slug'],
+		types: [undefined]
+	},
+	'blog/[[slug]]': {
+		pattern: /^\/blog(?:\/([^/]+))?\/?$/,
+		names: ['slug'],
+		types: [undefined]
+	},
+	'blog/[[slug=type]]/sub': {
+		pattern: /^\/blog(?:\/([^/]+))?\/sub\/?$/,
+		names: ['slug'],
+		types: ['type']
+	},
+	'blog/[[slug]].json': {
+		pattern: /^\/blog\/([^/]*)?\.json$/,
 		names: ['slug'],
 		types: [undefined]
 	},
@@ -62,6 +77,74 @@ for (const [key, expected] of Object.entries(tests)) {
 		assert.equal(actual.pattern.toString(), expected.pattern.toString());
 		assert.equal(actual.names, expected.names);
 		assert.equal(actual.types, expected.types);
+	});
+}
+
+const exec_tests = [
+	{
+		route: 'blog/[[slug]]/sub[[param]]',
+		path: '/blog/sub',
+		expected: { slug: '', param: '' }
+	},
+	{
+		route: 'blog/[[slug]]/sub[[param]]',
+		path: '/blog/slug/sub',
+		expected: { slug: 'slug', param: '' }
+	},
+	{
+		route: 'blog/[[slug]]/sub[[param]]',
+		path: '/blog/slug/subparam',
+		expected: { slug: 'slug', param: 'param' }
+	},
+	{
+		route: 'blog/[[slug]]/sub[[param]]',
+		path: '/blog/subparam',
+		expected: { slug: '', param: 'param' }
+	},
+	{
+		route: '[[slug]]/[...rest]',
+		path: '/slug/rest/sub',
+		expected: { slug: 'slug', rest: 'rest/sub' }
+	},
+	{
+		route: '[[slug]]/[...rest]',
+		path: '/slug/rest',
+		expected: { slug: 'slug', rest: 'rest' }
+	},
+	{
+		route: '[[slug]]/[...rest]',
+		path: '/slug',
+		expected: { slug: 'slug', rest: '' }
+	},
+	{
+		route: '[[slug]]/[...rest]',
+		path: '/',
+		expected: { slug: '', rest: '' }
+	},
+	{
+		route: '[...rest]/path',
+		path: '/rest/path',
+		expected: { rest: 'rest' }
+	},
+	{
+		route: '[[slug1]]/[[slug2]]',
+		path: '/slug1/slug2',
+		expected: { slug1: 'slug1', slug2: 'slug2' }
+	},
+	{
+		route: '[[slug1]]/[[slug2]]',
+		path: '/slug1',
+		expected: { slug1: 'slug1', slug2: '' }
+	}
+];
+
+for (const { path, route, expected } of exec_tests) {
+	test(`exec extracts params correctly for ${path}`, () => {
+		const { pattern, names, types } = parse_route_id(route);
+		const match = pattern.exec(path);
+		if (!match) throw new Error(`Failed to match ${path}`);
+		const actual = exec(match, names, types, {});
+		assert.equal(actual, expected);
 	});
 }