+++ packages/kit/test/apps/basics/src/routes/load/fetch-request-headers-invalid-access/+page.js
@@ -1,0 +1,8 @@
+/** @type {import('./$types').PageLoad} */
+export async function load({ fetch }) {
+	let res = await fetch('./irrelevant', {
+		method: 'GET'
+	});
+
+	res.headers.get('content-type');
+}