+++ packages/kit/test/apps/basics/src/routes/shadowed/redirect/[a]/+page.server.js
@@ -1,5 +1,5 @@
 import { redirect } from '@sveltejs/kit';
 
 export function load() {
-	throw redirect(302, '/shadowed/redirect/c');
+	throw redirect(302, '/shadowed/redirect/b');
 }